Task4[mandatory](oneill.com.)/Code_file/Task4/requirement(4).txt
attrs==20.2.0
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.1
itemloaders==1.0.3
jmespath==0.10.0
lxml==4.5.2
numpy==1.19.2
pandas==1.1.2
parsel==1.6.0
Protego==0.1.16
proxyscrape==0.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
queuelib==1.5.0
requests==2.24.0
schedule==0.6.0
scraperapi-sdk==0.2.2
Scrapy==2.3.0
scrapy-proxy-pool==0.1.9
scrapy-user-agents==0.1.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
Twisted==20.3.0
ua-parser==0.10.0
url-request-scheduler==0.3
urllib3==1.25.10
user-agents==2.2.0
w3lib==1.22.0
zope.interface==5.1.2
